+++ CHANGELOG.md
@@ -43,7 +43,7 @@
 * BUG修复：
   * [ ] 修复valueSetValue索引在索引时会被IDE认为栈溢出的问题，改为基于`ParadoxValueSetValuesSearch`和索引进行查找
 * 新增功能：
-  * [ ] 新增动作：导航到相关本地化和导航到相关图片（对于定义，在导航和右键菜单中，在导航到相关符号/Go to related symbol下面）
+  * [x] 新增动作：导航到相关本地化和导航到相关图片（对于定义，在导航和右键菜单中，在动作：导航到相关符号/Go to related symbol下面）
 * ［长期］完善CWT配置支持：
   * [ ] 优化：支持处理`value`和`value_set`自带的作用域信息（支持valueSetValueExpression，如`val@root.owner`）